+++ internal/core/domain/product.go
@@ -5,7 +5,7 @@ type ProductForm struct {
 	Name             string        `json:"name" validate:"required"`
 	Description      string        `json:"description" validate:"required"`
 	Category         string        `json:"category" validate:"required"`
-	Price            int           `json:"price" validate:"required"`
+	Price            float64       `json:"price" validate:"required"`
 	Images           []ProducImage `json:"images" validate:"required"`
 	ComboProductsIds *[]uint       `json:"comboProductsIds"`
 }
@@ -15,7 +15,7 @@ type ProductResponse struct {
 	Name          string             `json:"name" validate:"required"`
 	Description   string             `json:"description" validate:"required"`
 	Category      string             `json:"category" validate:"required"`
-	Price         int                `json:"price" validate:"required"`
+	Price         float64            `json:"price" validate:"required"`
 	Images        []ProducImage      `json:"images" validate:"required"`
 	ComboProducts *[]ProductResponse `json:"comboProducts"`
 }
@@ -29,17 +29,17 @@ type ProducImage struct {
 }
 
 type ComboForm struct {
-	Id          uint   `json:"id"`
-	Name        string `json:"name" validate:"required"`
-	Description string `json:"description" validate:"required"`
-	Price       int    `json:"price" validate:"required"`
-	Products    []uint `json:"products" validate:"required"`
+	Id          uint    `json:"id"`
+	Name        string  `json:"name" validate:"required"`
+	Description string  `json:"description" validate:"required"`
+	Price       float64 `json:"price" validate:"required"`
+	Products    []uint  `json:"products" validate:"required"`
 }
 
 type Combo struct {
 	Id          uint          `json:"id"`
 	Name        string        `json:"name"`
 	Description string        `json:"description"`
-	Price       int           `json:"price"`
+	Price       float64       `json:"price"`
 	Products    []ProductForm `json:"products"`
 }
